Refactor file paths for UI components
- Update import paths for `Copy` component in `pre.tsx`
- Update import paths for `Navbar` component in `layout.tsx`
- Update import paths for `Footer` component in `layout.tsx`
- Update import paths for `Logo` component in `logo.tsx`
- Update import paths for `PageBreadcrumb` component in `page.tsx`
- Update import paths for `Pagination` component in `page.tsx`
- Update import paths for `Toc` component in `page.tsx`
- Update import paths for `Feedback` component in `page.tsx`
- Update import paths for `BackToTop` component in `page.tsx`
- Remove unused `ThemeToggle` component
- Remove unused `Anchor` component
- Remove unused `PageMenu` component
- Update import paths for `Copy` component in `copy.tsx`
