+++ components/navigation/navbar.tsx
@@ -1,14 +1,13 @@
-import Image from "next/image";
 import Link from "next/link";
-import { Settings } from "@/settings/config";
+import { Logo } from "@/components/navigation/logo";
 import { LuArrowUpRight, LuGithub } from "react-icons/lu";
 
-import { ModeToggle } from "@/components/theme-toggle";
-import { SheetLeft } from "./sidebar";
-import { buttonVariants } from "./ui/button";
-import Search from "./search";
-import Anchor from "./anchor";
+import { ModeToggle } from "@/components/navigation/theme-toggle";
+import { SheetLeft } from "@/components/navigation/sidebar";
+import Search from "@/components/navigation/search";
+import Anchor from "@/components/navigation/anchor";
 import { Navigations, GitHubLink } from "@/settings/navigation";
+import { buttonVariants } from "@/components/ui/button";
 import { SheetClose } from "@/components/ui/sheet";
 
 export function Navbar() {
@@ -27,45 +26,25 @@ export function Navbar() {
           </div>
         </div>
 
-        <div className="flex items-center gap-3">
-          <div className="flex items-center gap-2">
-            <Search />
-            <div className="flex ml-2.5 sm:ml-0 gap-2">
+        <div className="flex items-center gap-2">
+          <Search />
+          <div className="flex sm:ml-0 gap-2">
             {GitHubLink.href && (
-              <>
-                <Link
-                  href={GitHubLink.href}
-                  className={buttonVariants({ variant: "ghost", size: "icon" })}
-                  target="_blank"
-                  rel="noopener noreferrer"
-                  aria-label="View the repository on GitHub"
-                >
-                  <LuGithub className="w-[1.1rem] h-[1.1rem]" />
-                </Link>
-              </>
+              <Link
+                href={GitHubLink.href}
+                className={buttonVariants({ variant: "outline", size: "icon" })}
+                target="_blank"
+                rel="noopener noreferrer"
+                aria-label="View the repository on GitHub"
+              >
+                <LuGithub className="w-[1.1rem] h-[1.1rem]" />
+              </Link>
             )}
-              <ModeToggle />
-            </div>
+            <ModeToggle />
           </div>
         </div>
       </div>
     </nav>
-  );
-}
-
-export function Logo() {
-  return (
-    <Link href="/" className="flex items-center gap-2.5">
-      <Image 
-        src={Settings.siteicon}
-        alt={`${Settings.title} main logo`}
-        width={34}
-        height={34}
-        loading="lazy"
-        decoding="async"
-      />
-      <h1 className="text-md font-semibold">{Settings.title}</h1>
-    </Link>
   );
 }
