+++ components/ui/copy.tsx
@@ -2,7 +2,7 @@
 
 import { useState } from "react";
 import { LuCheck, LuCopy } from "react-icons/lu";
-import { Button } from "./ui/button";
+import { Button } from "@/components/ui/button";
 
 import { cn } from "@/lib/utils";
